@@ -1,5 +1,5 @@
 import { Component, EventEmitter, Inject, Input, OnInit, Output } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { ApiResponseData, ApiResponseError, KnoraApiConnection, LoginResponse, LogoutResponse } from '@dasch-swiss/dsp-js';
 import { DspApiConnectionToken } from '../../../core/core.module';
 import { Session, SessionService } from '../../../core/session.service';
@@ -54,7 +54,7 @@ export class LoginFormComponent implements OnInit {
     session: Session;
 
     // form
-    form: FormGroup;
+    form: UntypedFormGroup;
 
     // show progress indicator
     loading = false;
@@ -103,7 +103,7 @@ export class LoginFormComponent implements OnInit {
         @Inject(DspApiConnectionToken) private _dspApiConnection: KnoraApiConnection,
         private _notification: NotificationService,
         private _sessionService: SessionService,
-        private _fb: FormBuilder
+        private _fb: UntypedFormBuilder
     ) { }
 
     ngOnInit() {
