@@ -3,7 +3,7 @@ import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ResourceAndPropertySelectionComponent } from './resource-and-property-selection.component';
 import { HarnessLoader } from '@angular/cdk/testing';
 import { Component, EventEmitter, Inject, Input, OnInit, Output, ViewChild } from '@angular/core';
-import { FormBuilder, FormGroup, ReactiveFormsModule } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, ReactiveFormsModule } from '@angular/forms';
 import { DspApiConnectionToken } from '../../../core';
 import { OntologyCache } from '@dasch-swiss/dsp-js/src/cache/ontology-cache/OntologyCache';
 import { MockOntology, ReadOntology, ResourceClassDefinition, ResourcePropertyDefinition } from '@dasch-swiss/dsp-js';
@@ -22,7 +22,7 @@ import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 })
 class TestSelectResourceClassComponent {
 
-    @Input() formGroup: FormGroup;
+    @Input() formGroup: UntypedFormGroup;
 
     @Input() resourceClassDefinitions: ResourceClassDefinition[];
 
@@ -40,7 +40,7 @@ class TestSelectResourceClassComponent {
 class TestSelectPropertyComponent {
 
     // parent FormGroup
-    @Input() formGroup: FormGroup;
+    @Input() formGroup: UntypedFormGroup;
 
     // index of the given property (unique)
     @Input() index: number;
@@ -66,13 +66,13 @@ class TestHostComponent implements OnInit {
 
     @ViewChild('resClassAndProp') resourceClassAndPropertySelection: ResourceAndPropertySelectionComponent;
 
-    form: FormGroup;
+    form: UntypedFormGroup;
 
     activeOntology = 'http://0.0.0.0:3333/ontology/0001/anything/v2';
 
     restrictByResourceClass?: string;
 
-    constructor(@Inject(FormBuilder) private _fb: FormBuilder) {
+    constructor(@Inject(UntypedFormBuilder) private _fb: UntypedFormBuilder) {
     }
 
     ngOnInit() {
