@@ -1,5 +1,5 @@
 import { Component, Inject, Input, OnInit, OnDestroy } from '@angular/core';
-import { FormBuilder, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 import { Constants } from '@dasch-swiss/dsp-js';
 import { CustomRegex } from '../../../../../../viewer/values/custom-regex';
 import { PropertyValue, Value, ValueLiteral } from '../operator';
@@ -15,13 +15,13 @@ const resolvedPromise = Promise.resolve(null);
 export class SearchUriValueComponent implements OnInit, OnDestroy, PropertyValue {
 
     // parent FormGroup
-    @Input() formGroup: FormGroup;
+    @Input() formGroup: UntypedFormGroup;
 
     type = Constants.UriValue;
 
-    form: FormGroup;
+    form: UntypedFormGroup;
 
-    constructor(@Inject(FormBuilder) private _fb: FormBuilder) {
+    constructor(@Inject(UntypedFormBuilder) private _fb: UntypedFormBuilder) {
 
     }
 
