@@ -1,5 +1,5 @@
 import { Component, Inject, Input, OnChanges, OnDestroy, OnInit, SimpleChanges, ViewChild } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { CreateColorValue, ReadColorValue, UpdateColorValue } from '@dasch-swiss/dsp-js';
 import { Subscription } from 'rxjs';
 import { BaseValueComponent } from '../base-value.component';
@@ -23,15 +23,15 @@ export class ColorValueComponent extends BaseValueComponent implements OnInit, O
 
     @Input() showHexCode = false;
 
-    valueFormControl: FormControl;
-    commentFormControl: FormControl;
-    form: FormGroup;
+    valueFormControl: UntypedFormControl;
+    commentFormControl: UntypedFormControl;
+    form: UntypedFormGroup;
     valueChangesSubscription: Subscription;
     customValidators = [Validators.pattern(CustomRegex.COLOR_REGEX)];
     matcher = new ValueErrorStateMatcher();
     textColor: string;
 
-    constructor(@Inject(FormBuilder) private _fb: FormBuilder) {
+    constructor(@Inject(UntypedFormBuilder) private _fb: UntypedFormBuilder) {
         super();
     }
 
@@ -46,8 +46,8 @@ export class ColorValueComponent extends BaseValueComponent implements OnInit, O
     ngOnInit() {
 
         // initialize form control elements
-        this.valueFormControl = new FormControl(null);
-        this.commentFormControl = new FormControl(null);
+        this.valueFormControl = new UntypedFormControl(null);
+        this.commentFormControl = new UntypedFormControl(null);
 
         // subscribe to any change on the comment and recheck validity
         this.valueChangesSubscription = this.commentFormControl.valueChanges.subscribe(
