@@ -1,6 +1,6 @@
 import { Component, DoCheck, ElementRef, HostBinding, Input, OnDestroy, OnInit, Optional, Self } from '@angular/core';
 import { MatFormFieldControl } from '@angular/material/form-field';
-import { AbstractControl, ControlValueAccessor, FormBuilder, FormControl, FormGroup, FormGroupDirective, NgControl, NgForm, ValidatorFn, Validators } from '@angular/forms';
+import { AbstractControl, ControlValueAccessor, UntypedFormBuilder, UntypedFormControl, UntypedFormGroup, FormGroupDirective, NgControl, NgForm, ValidatorFn, Validators } from '@angular/forms';
 import { Subject } from 'rxjs';
 import { FocusMonitor } from '@angular/cdk/a11y';
 import { coerceBooleanProperty } from '@angular/cdk/coercion';
@@ -21,14 +21,14 @@ export class Interval {
 
 /** Error when invalid control is dirty, touched, or submitted. */
 export class IntervalInputErrorStateMatcher implements ErrorStateMatcher {
-    isErrorState(control: FormControl | null, form: FormGroupDirective | NgForm | null): boolean {
+    isErrorState(control: UntypedFormControl | null, form: FormGroupDirective | NgForm | null): boolean {
         const isSubmitted = form && form.submitted;
         return !!(control && control.invalid && (control.dirty || control.touched || isSubmitted));
     }
 }
 
 /** Interval must have a start and end of the same type, either both numbers or both null */
-export function startEndSameTypeValidator(otherInterval: FormControl): ValidatorFn {
+export function startEndSameTypeValidator(otherInterval: UntypedFormControl): ValidatorFn {
     return (control: AbstractControl): { [key: string]: any } | null => {
 
         // valid if both start and end are null or have values
@@ -39,6 +39,7 @@ export function startEndSameTypeValidator(otherInterval: FormControl): Validator
 }
 
 class MatInputBase {
+    readonly stateChanges = new Subject<void>();
     constructor(public _defaultErrorStateMatcher: ErrorStateMatcher,
         public _parentForm: NgForm,
         public _parentFormGroup: FormGroupDirective,
@@ -57,7 +58,7 @@ const _MatInputMixinBase =
 export class IntervalInputComponent extends _MatInputMixinBase implements ControlValueAccessor, MatFormFieldControl<Interval>, DoCheck, CanUpdateErrorState, OnDestroy, OnInit {
     static nextId = 0;
 
-    form: FormGroup;
+    form: UntypedFormGroup;
     stateChanges = new Subject<void>();
     @HostBinding() id = `dsp-interval-input-${IntervalInputComponent.nextId++}`;
     focused = false;
@@ -65,8 +66,8 @@ export class IntervalInputComponent extends _MatInputMixinBase implements Contro
     controlType = 'dsp-interval-input';
     matcher = new IntervalInputErrorStateMatcher();
 
-    startIntervalControl: FormControl;
-    endIntervalControl: FormControl;
+    startIntervalControl: UntypedFormControl;
+    endIntervalControl: UntypedFormControl;
 
     @Input() intervalStartLabel = 'start';
     @Input() intervalEndLabel = 'end';
@@ -152,7 +153,7 @@ export class IntervalInputComponent extends _MatInputMixinBase implements Contro
 
     @Input() errorStateMatcher: ErrorStateMatcher;
 
-    constructor(fb: FormBuilder,
+    constructor(fb: UntypedFormBuilder,
         @Optional() @Self() public ngControl: NgControl,
         private _fm: FocusMonitor,
         private _elRef: ElementRef<HTMLElement>,
@@ -162,8 +163,8 @@ export class IntervalInputComponent extends _MatInputMixinBase implements Contro
 
         super(_defaultErrorStateMatcher, _parentForm, _parentFormGroup, ngControl);
 
-        this.startIntervalControl = new FormControl(null);
-        this.endIntervalControl = new FormControl(null);
+        this.startIntervalControl = new UntypedFormControl(null);
+        this.endIntervalControl = new UntypedFormControl(null);
 
         this.form = fb.group({
             start: this.startIntervalControl,
