@@ -3,9 +3,9 @@ import { BaseValueComponent } from '../base-value.component';
 import { ValueErrorStateMatcher } from '../value-error-state-matcher';
 import { CreateIntervalValue, ReadIntervalValue, UpdateIntervalValue } from '@dasch-swiss/dsp-js';
 import {
-    FormBuilder,
-    FormControl,
-    FormGroup
+    UntypedFormBuilder,
+    UntypedFormControl,
+    UntypedFormGroup
 } from '@angular/forms';
 import { Subscription } from 'rxjs';
 import { Interval, IntervalInputComponent } from './interval-input/interval-input.component';
@@ -24,10 +24,10 @@ export class IntervalValueComponent extends BaseValueComponent implements OnInit
 
     @Input() displayValue?: ReadIntervalValue;
 
-    valueFormControl: FormControl;
-    commentFormControl: FormControl;
+    valueFormControl: UntypedFormControl;
+    commentFormControl: UntypedFormControl;
 
-    form: FormGroup;
+    form: UntypedFormGroup;
 
     valueChangesSubscription: Subscription;
 
@@ -35,7 +35,7 @@ export class IntervalValueComponent extends BaseValueComponent implements OnInit
 
     matcher = new ValueErrorStateMatcher();
 
-    constructor(@Inject(FormBuilder) private _fb: FormBuilder) {
+    constructor(@Inject(UntypedFormBuilder) private _fb: UntypedFormBuilder) {
         super();
     }
 
@@ -53,9 +53,9 @@ export class IntervalValueComponent extends BaseValueComponent implements OnInit
 
     ngOnInit() {
         // initialize form control elements
-        this.valueFormControl = new FormControl(null);
+        this.valueFormControl = new UntypedFormControl(null);
 
-        this.commentFormControl = new FormControl(null);
+        this.commentFormControl = new UntypedFormControl(null);
 
         // subscribe to any change on the comment and recheck validity
         this.valueChangesSubscription = this.commentFormControl.valueChanges.subscribe(
