@@ -9,7 +9,7 @@ import {
     Output,
     SimpleChanges
 } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 import { Constants, CreateTextValueAsXml, ReadTextValueAsXml, UpdateTextValueAsXml } from '@dasch-swiss/dsp-js';
 import * as Editor from 'ckeditor5-custom-build';
 import { Subscription } from 'rxjs';
@@ -34,10 +34,10 @@ export class TextValueAsXMLComponent extends BaseValueComponent implements OnIni
 
     @Output() internalLinkHovered: EventEmitter<string> = new EventEmitter<string>();
 
-    valueFormControl: FormControl;
-    commentFormControl: FormControl;
+    valueFormControl: UntypedFormControl;
+    commentFormControl: UntypedFormControl;
 
-    form: FormGroup;
+    form: UntypedFormGroup;
 
     valueChangesSubscription: Subscription;
     matcher = new ValueErrorStateMatcher();
@@ -62,7 +62,7 @@ export class TextValueAsXMLComponent extends BaseValueComponent implements OnIni
     // TODO: get this from config via AppInitService
     readonly resourceBasePath = 'http://rdfh.ch/';
 
-    constructor(@Inject(FormBuilder) private fb: FormBuilder) {
+    constructor(@Inject(UntypedFormBuilder) private fb: UntypedFormBuilder) {
         super();
     }
 
@@ -127,9 +127,9 @@ export class TextValueAsXMLComponent extends BaseValueComponent implements OnIni
         };
 
         // initialize form control elements
-        this.valueFormControl = new FormControl(null);
+        this.valueFormControl = new UntypedFormControl(null);
 
-        this.commentFormControl = new FormControl(null);
+        this.commentFormControl = new UntypedFormControl(null);
 
         this.valueChangesSubscription = this.commentFormControl.valueChanges.subscribe(
             data => {
